requirements.txt
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
bcrypt==4.0.1
billiard==4.2.0
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comtypes==1.2.0
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
decorator==5.1.1
Django==4.2.7
djangorestframework==3.14.0
executing==2.0.1
fonttools==4.44.1
getmac==0.9.4
idna==3.4
ipython==8.17.2
jedi==0.19.1
joblib==1.3.2
kiwisolver==1.4.5
kombu==5.3.4
matplotlib==3.8.1
matplotlib-inline==0.1.6
numpy==1.26.0
optional-django==0.3.0
packaging==23.2
pandas==2.1.3
paramiko==3.3.1
parso==0.8.3
Pillow==10.1.0
ping3==4.0.4
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
pygame==2.5.2
Pygments==2.16.1
PyNaCl==1.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywifi==1.1.12
PyX==0.16
requests==2.31.0
scapy==2.5.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
speedtest-cli==2.1.3
sqlparse==0.4.4
stack-data==0.6.3
threadpoolctl==3.2.0
traitlets==5.13.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.10
webpack==6.0.0
who-is-on-my-wifi==1.3.5
